requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-crispy-forms==1.7.2
django-filter==2.2.0
django-mathfilters==0.4.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
idna==2.8
Markdown==3.1.1
Pillow==6.1.0
PyJWT==1.7.1
python-http-client==3.1.0
pytz==2019.1
requests==2.22.0
sendgrid==6.0.5
sqlparse==0.3.0
urllib3==1.25.3
